refactor(sdk/metric): remove redundant int32 conversion for scale bounds check

+++ sdk/metric/internal/aggregate/exponential_histogram.go
@@ -201,7 +201,7 @@ func (p *expoHistogramDataPoint[N]) scaleChange(bin, startBin int32, length int)
 		low >>= 1
 		high >>= 1
 		count++
-		if count > int32(expoMaxScale-expoMinScale) {
+		if count > expoMaxScale-expoMinScale {
 			return count
 		}
 	}
